% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GSU.R
\name{cond.GSU}
\alias{cond.GSU}
\title{Condition an operating model for Green Sea Urchin}
\usage{
cond.GSU(
  RCMinput,
  sims = 12,
  max_F = 3,
  comp_like = "multinomial",
  resample = T,
  parallel = T,
  silent = T
)
}
\arguments{
\item{sims}{Integer or vector of integers - the number of simulations for the operating model (e.g. 96) or the specific vector of simulations for the operating model (e.g. 13:48)}

\item{max_F}{Positive real number - the maximum instantaneous mortality rate in any historical time step}

\item{comp_like}{The likelihood function used for composition data (age / length) c("multinomial", "lognormal", "mvlogistic", "dirmult1", "dirmult2") see ?RCM}

\item{resample}{Logical - should parameters be drawn from the variance covariance matrix of the MLE fit (T) or just the MLE parameter values (F)}

\item{parallel}{Logical - should model fitting use multiple cores}

\item{silent}{Logical - should all RCM messages be repressed (ie not read out onto the console)?}

\item{In}{A list object of class ('In') that includes the operating model parameters (slot 2, class OM) and data (slot 3, class RCMinput) for conditioning}
}
\description{
Uses the RCM model of OpenMSE to fit an operating model to data
}
\examples{
cond.GSU("In.GSU.12")
}
\seealso{
\link{RCM}
}
\author{
T. Carruthers
}
