% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_GSU.R
\name{Obs_GSU}
\alias{Obs_GSU}
\title{Observation error specification for Green Sea Urchin}
\usage{
Obs_GSU(Hist, Catch_CV = 0.01, Catch_bias = 1, Index_CV = 0.1, Index_bias = 1)
}
\arguments{
\item{Hist}{Object of class 'Hist' made by runMSE(OM, Hist=T)}

\item{Catch_CV}{Positive real number - the coeficient of variation of catch observations}

\item{Catch_bias}{Positive real number - the mean bias of the catch observations. 1 is unbiased, 1.1 is a 10 percent positive bias on average}

\item{Index_CV}{Positive real number - the coeficient of variation of relative or absolute abundance observations}

\item{Index_bias}{Positive real number - the mean bias of the relatie or absolute abundance observations. 1 is unbiased, 1.1 is a 10 percent positive bias on average}
}
\description{
Controls bias and imprecision of observed catches and indices
}
\author{
T. Carruthers
}
